requirements.txt
requests~=2.31.0
pandas~=1.3.5
pycryptodome~=3.19.0
ddddocr~=1.4.8
ruamel.yaml~=0.17.35
lxml~=5.4.0